refact: move api-requests to burger-api.js, refact inline styles, refact func calcSum
