+++ src/components/Burger-Constructor/Burger-Constructor.module.css
@@ -5,6 +5,12 @@
     margin-bottom: 40px;    
 }
 
+.container {
+    display: flex;
+    flex-direction: column;
+    padding: 0 16px;
+}
+
 .list {    
     list-style-type: none;
     margin: 0;
@@ -28,9 +34,22 @@
     box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
 }
 
+.item {
+    display: flex;
+    column-gap: 8px;
+    align-items: center;
+    padding-bottom: 16px;
+}
+
 .total {
     display: flex;
     justify-content: end;
     align-items: center;
     padding: 40px 40px 0 0;
+}
+
+.order {
+    transform: scale(1.5);
+    z-index: -1;
+    margin-right: 40px;
 }
