--- src/global/global.js
@@ -1,2 +1,0 @@
-export const url = 'https://norma.nomoreparties.space/api/';
-
